fix: fix double free and add test
