+++ src/rust/src/lib.rs
@@ -1,5 +1,6 @@
 use extendr_api::prelude::*;
 
+pub(crate) mod utils;
 mod kractor; // kraken extractor
 
 // https://extendr.github.io/extendr/extendr_api/#returning-resultt-e-to-r
